refactor(resources): split resource payload modules
