+++ crates/codelens-mcp/src/main.rs
@@ -8,7 +8,10 @@ mod error;
 mod mutation_gate;
 mod prompts;
 mod protocol;
+mod resource_analysis;
+mod resource_catalog;
 mod resource_context;
+mod resource_profiles;
 mod resources;
 mod runtime_types;
 mod server;
